added most recent history item to status
