+++ components/BillStatus/BillStatus.js
@@ -1,6 +1,8 @@
 import React, {useState} from "react";
 import { Button, Modal } from 'react-bootstrap'
 import BillHistory from "../BillHistory/BillHistory";
+import BillHistoryTable from "../BillHistoryTable/BillHistoryTable";
+import { documentHistoryActions } from "../MockAPIResponseDocumentHistoryActions";
 
 const BillStatus = (props) => {
   const bill = props.bill
@@ -10,26 +12,27 @@ const BillStatus = (props) => {
   const handleCloseBillStatus = () => setShowBillStatus(false);
 
     return (
-  <>
-      <Button variant="primary" className="m-1" onClick={handleShowBillStatus}>
-        Status
-      </Button>
-      <Modal show={showBillStatus} onHide={handleCloseBillStatus} size="lg">
-        <Modal.Header closeButton onClick={handleCloseBillStatus}>
-            <Modal.Title>{bill ? bill.BillNumber : ""}</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          <>
-            <div className="text-center">
-              Bill Status
-            </div>
-            <div className=" d-flex justify-content-center">
-              <BillHistory bill={bill} />
-            </div>
-          </>
-        </Modal.Body>
-      </Modal>
-    </>
+      <>
+        <Button variant="primary" className="m-1" onClick={handleShowBillStatus}>
+          Status
+        </Button>
+        <Modal show={showBillStatus} onHide={handleCloseBillStatus} size="lg">
+          <Modal.Header closeButton onClick={handleCloseBillStatus}>
+              <Modal.Title>{bill ? bill.BillNumber : ""}</Modal.Title>
+          </Modal.Header>
+          <Modal.Body>
+            <>
+              <div className="text-center">
+                Bill Status
+              </div>
+              <BillHistoryTable documentHistoryActions={[documentHistoryActions[documentHistoryActions.length - 1]]}/>
+              <div className=" d-flex justify-content-center">
+                <BillHistory bill={bill} />
+              </div>
+            </>
+          </Modal.Body>
+        </Modal>
+      </>
     )
 }
